profiles
turn and vel profiles are defined. For simplicity in this first iteration they are rectangular. In order to first try to check if everything is working. The robot moves from point to point and turns, All this at a costant velocity.

comments:
dst is not a tuple is an dst.   got:invalid index to scalar variable when doing dst[0]. Solved